indexからの解放
