--- src/services/subscriptions/index.ts
@@ -1,2 +1,0 @@
-export * from './subscribe';
-export * from './verify';
